Exclude admin routes from Google Analytics tracking
Admin pageviews were polluting advertising analytics because GA loaded
unconditionally for all routes. Moves GA initialization into a client
component that returns null on any /admin/* pathname.
